requirements.txt
streamlit
faiss-cpu
sentence-transformers
pymupdf
camelot-py[cv]
groq
numpy
pandas
python-dotenv